refactor: move `postcss` plugin into own file
